refactor(politics-tracker): move function from `politics-form` to `utils`
